creates a dedicated namespace for test classes and move the TestDbEntity class to a Mocks directory as it is a custom mocking class
